requirements.txt
nba_api==1.4.1
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0
supabase==2.3.0
python-dotenv==1.0.0
